move storage stuff into single file
